Add benchmarking module with signal handling and dynamic worker management; update CSV processing and results handling
