+++ results/results_9.csv
@@ -1482,15 +1482,15 @@ Block Sort,9,480,0.00002000
 Block Sort,9,481,0.00002730
 Block Sort,9,482,0.00001900
 Block Sort,9,483,0.00002710
-Block Sort,9,484,0.00001890
-Block Sort,9,485,0.00002910
-Block Sort,9,486,0.00002040
-Block Sort,9,487,0.00001930
-Block Sort,9,488,0.00001960
-Block Sort,9,489,0.00002470
-Block Sort,9,490,0.00002350
-Block Sort,9,491,0.00001960
-Block Sort,9,492,0.00001950
+Block Sort,9,484,0.00002890
+Block Sort,9,485,0.00001810
+Block Sort,9,486,0.00001520
+Block Sort,9,487,0.00001410
+Block Sort,9,488,0.00001310
+Block Sort,9,489,0.00001380
+Block Sort,9,490,0.00001310
+Block Sort,9,491,0.00001460
+Block Sort,9,492,0.00002160
 Block Sort,9,493,0.00002360
 Block Sort,9,494,0.00001880
 Block Sort,9,495,0.00002420
